Camera3: Set sync frame number to int64
Even though frame number in setting is unsigned 32bit integer,
the sync frame number in capture result is int64 to accommodate
negative values.

+++ camera/QCamera2/HAL3/QCamera3HWI.cpp
@@ -1961,7 +1961,7 @@ QCamera3HardwareInterface::translateFromHalMetadata(
     camMetadata.update(ANDROID_REQUEST_ID, &request_id, 1);
 
     if (IS_META_AVAILABLE(CAM_INTF_META_FRAME_NUMBER, metadata)) {
-        int32_t frame_number = *(uint32_t *) POINTER_OF_META(CAM_INTF_META_FRAME_NUMBER, metadata);
+        int64_t frame_number = *(uint32_t *) POINTER_OF_META(CAM_INTF_META_FRAME_NUMBER, metadata);
         camMetadata.update(ANDROID_SYNC_FRAME_NUMBER, &frame_number, 1);
     }
